# Conflicts:
#	redgamejam/Assets/Scenes/Cave.unity
#	redgamejam/Assets/Scripts/ScoreManager.cs
#	redgamejam/Logs/AssetImportWorker0-prev.log
#	redgamejam/Logs/AssetImportWorker0.log
#	redgamejam/Logs/AssetImportWorker1-prev.log
#	redgamejam/Logs/AssetImportWorker1.log
#	redgamejam/Logs/AssetImportWorker2-prev.log
#	redgamejam/Logs/shadercompiler-AssetImportWorker0.log
#	redgamejam/UserSettings/Layouts/default-2021.dwlt

+++ redgamejam/Assets/Scripts/Tile.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Tile : MonoBehaviour
 {
     public static Tile instance;
-    public int moveTimer;
+    public float moveTimer;
     public int maxHP = 1;
     public int currentHP;
     private Camera mainCamera;
